[packaging] Move param related functionality into its own package
