+++ src/test/java/at/fhhagenberg/sqe/sqelevator/mock/Launcher.java
@@ -1,7 +1,5 @@
 package at.fhhagenberg.sqe.sqelevator.mock;
 
-import at.fhhagenberg.sqe.ui.ElevatorParams;
-import at.fhhagenberg.sqe.ui.ParamUtils;
 import javafx.application.Application;
 
 public class Launcher {
